Use the displayName for stacks instead of the stack name
Later versions of the CDK CLI warn you not to use the stack name, and to
instead use a "path-like" name to identify stacks that you supply to the
command line.

For example, if you have:
```
Stack1:
  - Stack2
```
And you run:
```
cdk deploy Stack2
```
You will see:
```
Selecting stack by identifier "Stack2". This identifier is deprecated and will be removed in v2. Please use "Stack1/Stack2" instead.
```

This commit changes this libraries behaviour to use the "path-like"
identifiers.
If the `"displayName"` attribute is not present (eg. older versions of
the CDK CLI), then we fall back to the stack name.