Prevent CDK from trying to deploy dependencies on it's own.
Because this library manages all of that on it's own, we don't need CDK
to deploy dependencies. If it tries then errors occur due to stacks
being updated multiple times (`AlreadyExistsException: ChangeSet
cdk-deploy-change-set cannot be created due to a mismatch with existing
attribute Description`).

Luckily, we can use the `--exclusive` flag on the `cdk` command to tell
it to only deploy the stack we specify.